ai-operator/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
httpx
respx


https://huggingface.co/meta-llama/Llama-3.1-8B/tree/main